Expose the error message when the TP-Link API returns an unknown error
This improves the handling of unknown errors returned by TPLink,
including the error message in the thrown error as well as the
error code.

Just seeing a code is rather opaque and you end up having to add
debugging code into the library to actually see what came back.
Exposing the message avoids this.