Generate a random terminal UUID when starting a `cloudLogin` session
When you call `cloudLogin` to log in to the TP-Link API, you have
to send a `terminalUUID`. At the moment, this is hard-coded, and
is triggering a `-20004` error:

```json
{"error_code":-20004,"msg":"API rate limit exceeded"}
```

This switches to generate a new terminal UUID on each call using
the `uuid` npm package, which gets around this rate limiting
issue.

Fixes #22.